profit

+++ src/des/seguimiento3/estadoDelSistema/Producto.java
@@ -1,20 +1,20 @@
 package seguimiento3.estadoDelSistema;
 
 public enum Producto{
-    BEBIDA_SALUDABLE("Bebida saludable",600.0,1200.0),
-    PANADERIA("Panaderia",400.0,850.0);
+    BEBIDA_SALUDABLE("Bebida saludable",600,1200),
+    PANADERIA("Panaderia",400,850);
 
     private final String nombre;
-    private final Double costoUnitario;
-    private final Double precioUnitario;
+    private final Integer costoUnitario;
+    private final Integer precioUnitario;
 
-    private Producto(String nombre, Double costoUnitario, Double precioUnitario){
+    private Producto(String nombre, Integer costoUnitario, Integer precioUnitario){
         this.nombre = nombre;
         this.costoUnitario = costoUnitario;
         this.precioUnitario = precioUnitario;
     }
 
-    public Double getBeneficioUnitario(){
+    public Integer getBeneficioUnitario(){
         return this.precioUnitario - this.costoUnitario;
     }
